# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-floatpower",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute `base.powf(exp)`.",
        "description": "Input validation is delegated to `srvcs-isnumber` over HTTP (the single\nsource of truth for \"is this a number\"), once per operand. Both integer and\nfloat inputs are accepted — this is a float service. If the dependency is\nunreachable, this service reports itself degraded rather than guessing. If\nthe computed result is not a real number (`NaN`, e.g. a negative base with a\nfractional exponent), it is rejected as a domain error.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PowerResponse"
                }
              }
            }
          },
          "422": {
            "description": "an operand is not a number, or the result is not a real number"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "base",
          "exp"
        ],
        "properties": {
          "base": {
            "type": "object"
          },
          "exp": {
            "type": "object"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "PowerResponse": {
        "type": "object",
        "required": [
          "base",
          "exp",
          "result"
        ],
        "properties": {
          "base": {
            "type": "object"
          },
          "exp": {
            "type": "object"
          },
          "result": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}